Make openid-connect-configuration.php more robust  * Handle keys without the option x5c field. If these are RSA keys they can be parsed if phpseclib is available, otherwise they are ignored.  * log warnings and errors to stderr using error_log.  * Improve user feedback on issues
This uses phpseclib/phpseclib and requires dev-master to work. This
isn't added as a dependency to composer.json because it isn't run
during normal operation, and finding a set of dependencies that
includes both simplesaml and phpseclib which allows both to be
installed at the same time proved to be quite tricky.